Move players list state to BoardGameContext
