+++ src/games/kalambury/board/Board.stories.js
@@ -1,6 +1,4 @@
 import React from "react";
-import { action } from "@storybook/addon-actions";
-import { boolean, number, text, select } from "@storybook/addon-knobs";
 
 import Board from "./Board";
