+++ src/games/kalambury/board/GameBoard.js
@@ -3,12 +3,14 @@ import { Segment, Button, Header } from "semantic-ui-react";
 import Confetti from "react-dom-confetti";
 import { shuffle } from "lodash";
 import { useTranslation } from "react-i18next";
+
 import { timerFormat, currentTime } from "../utils/time";
-import DrawArea from "../DrawArea";
-import GuessingBoard from "./GuessingBoard";
 import { COLORS } from "config/constants";
 import { useBoardGame } from "contexts/BoardGameContext";
 import filterActions from "utils/user/filterActions";
+
+import DrawArea from "../DrawArea";
+import GuessingBoard from "./GuessingBoard";
 
 const confettiConfig = {
   angle: 90,
@@ -23,35 +25,33 @@ const confettiConfig = {
   colors: shuffle(COLORS).slice(0, 5),
 };
 
-const GameBoard = ({
-  envokeLastAnswer,
-  guessInputRef,
-  guess,
-  setGuess,
-  canChangePhrase,
-  ChangePhrase,
-}) => {
-  const { G, ctx, moves, playerID, rawClient } = useBoardGame();
+const GameBoard = ({ guess, setGuess, envokeLastAnswer, guessInputRef }) => {
+  const {
+    G: { turnEndTime, canChangePhrase, actions },
+    moves: { NotifyTimeout, ChangePhrase, UpdateDrawing },
+    player: { isDrawing },
+    playerID,
+    rawClient,
+  } = useBoardGame();
   const { t } = useTranslation("kalambury");
   const [lines, setLines] = useState([]);
-  const isDrawing = ctx.activePlayers[playerID] === "draw";
-  const [remainingSeconds, setRemainingSeconds] = useState(G.turnEndTime - currentTime());
+  const [remainingSeconds, setRemainingSeconds] = useState(turnEndTime - currentTime());
   const [lastSuccess, setLastSuccess] = useState(false);
-  const lastUserGuess = filterActions(G.actions, playerID, "guess")[0] || {
+  const lastUserGuess = filterActions(actions, playerID, "guess")[0] || {
     id: null,
     success: false,
   };
 
   useEffect(() => {
     const interval = setInterval(() => {
-      const seconds = G.turnEndTime - currentTime();
+      const seconds = turnEndTime - currentTime();
       if ((seconds <= 0 && isDrawing) || seconds < 5) {
-        moves.NotifyTimeout();
+        NotifyTimeout();
       }
       setRemainingSeconds(Math.max(seconds, 0));
     }, 1000);
     return () => clearInterval(interval);
-  }, [G.turnEndTime, setRemainingSeconds, moves, isDrawing]);
+  }, [turnEndTime, setRemainingSeconds, NotifyTimeout, isDrawing]);
 
   useEffect(() => {
     setLastSuccess(lastUserGuess.success);
@@ -77,7 +77,7 @@ const GameBoard = ({
   }, [isDrawing, rawClient.transport.socket]);
 
   useEffect(() => {
-    if (isDrawing) moves.UpdateDrawing(lines);
+    if (isDrawing) UpdateDrawing(lines);
   }, [isDrawing, moves, lines]);
 
   const handleChangePhrase = () => {
