+++ src/games/kalambury/board/WaitingBoard.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { action } from "@storybook/addon-actions";
-import { withKnobs, boolean, text } from "@storybook/addon-knobs";
+import { boolean, text } from "@storybook/addon-knobs";
 
 import WaitingBoard from "./WaitingBoard";
